Added "beds" to second level
Determined that it would be cool to base ech level off of a dorm housing complex, wiht the first being a cafeteria, the second being bedrooms, and the third being a common area. I added bed-like objects to the second level, added a jump onto a ramping platform off o a bed, and some cover as "separation" between the "rooms" and the rest of the building. I also added some needed basic changes.