Many many Changes
Added a drop-down area on the top level in two corners and four "table" structures on either side. Added more cover to the second level and spaced out the beds a bit more, as well as adding a bed at the bottom of the third floor drop-downs. In order to create a bit more flow and variety to the third flow, I carved out a little archway nook in the two pillars and added an aromor pickup. I also swapped the armor in the elevator sections with rockets in order to encourage players who grab the rocket launcher to move down and play in more areas of the map. Based off of some playtest feedback, I created a secondary ramp and jump on the side of the weapon pedestals present on every level, and replaced the stairs on the bottom level alcoves with ramps to encourage faster player movement. Lastly, in order to start experimenting with textures and shaders, I used the xatrix WAD pack (found on quaddicted.com/files/wads) on all of my brushes, and added lights to ech floor. I'm not super happy with how the textures work for my theme, and may try downloading a few other WAD packs to see if I can get anything better looking. Lighting is its own beast and while I have a different color for each level, I'm still working on the delay, wait, intensity, and placememtn of each.